Move checkpoint io plugins from pl/plugins/io to lite/plugins/io (#14519)
