+++ src/lightning_lite/plugins/io/checkpoint_plugin.py
