+++ src/lightning_lite/plugins/io/torch_plugin.py
@@ -16,11 +16,11 @@ import os
 from typing import Any, Callable, Dict, Optional
 
 import pytorch_lightning as pl
+from lightning_lite.plugins.io.checkpoint_plugin import CheckpointIO
 from lightning_lite.utilities.cloud_io import atomic_save, get_filesystem
 from lightning_lite.utilities.cloud_io import load as pl_load
+from lightning_lite.utilities.rank_zero import rank_zero_warn
 from lightning_lite.utilities.types import _PATH
-from pytorch_lightning.plugins.io.checkpoint_plugin import CheckpointIO
-from pytorch_lightning.utilities.rank_zero import rank_zero_warn
 
 log = logging.getLogger(__name__)
