+++ src/pytorch_lightning/plugins/io/hpu_plugin.py
@@ -17,10 +17,10 @@ from typing import Any, Dict, Optional
 
 import torch
 
+from lightning_lite.plugins.io.torch_plugin import TorchCheckpointIO
 from lightning_lite.utilities.apply_func import move_data_to_device
 from lightning_lite.utilities.cloud_io import atomic_save, get_filesystem
 from lightning_lite.utilities.types import _PATH
-from pytorch_lightning.plugins.io.torch_plugin import TorchCheckpointIO
 
 
 class HPUCheckpointIO(TorchCheckpointIO):
