+++ src/pytorch_lightning/strategies/parallel.py
@@ -20,6 +20,7 @@ from torch import Tensor
 
 import pytorch_lightning as pl
 from lightning_lite.plugins.environments.cluster_environment import ClusterEnvironment
+from lightning_lite.plugins.io.checkpoint_plugin import CheckpointIO
 from lightning_lite.utilities.distributed import (
     _get_process_group_backend_from_env,
     all_gather_ddp_if_available,
@@ -27,7 +28,6 @@ from lightning_lite.utilities.distributed import (
     ReduceOp,
 )
 from pytorch_lightning.plugins import LayerSync
-from pytorch_lightning.plugins.io.checkpoint_plugin import CheckpointIO
 from pytorch_lightning.plugins.precision import PrecisionPlugin
 from pytorch_lightning.strategies.strategy import Strategy
 from pytorch_lightning.utilities.rank_zero import rank_zero_deprecation
