+++ tests/tests_pytorch/accelerators/test_cpu.py
@@ -6,10 +6,10 @@ import pytest
 import torch
 
 import pytorch_lightning as pl
+from lightning_lite.plugins.io.torch_plugin import TorchCheckpointIO
 from pytorch_lightning import Trainer
 from pytorch_lightning.accelerators import CPUAccelerator
 from pytorch_lightning.demos.boring_classes import BoringModel
-from pytorch_lightning.plugins.io.torch_plugin import TorchCheckpointIO
 from pytorch_lightning.plugins.precision.precision_plugin import PrecisionPlugin
 from pytorch_lightning.strategies import SingleDeviceStrategy
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
